+++ modules/core-config/local.tf
@@ -26,12 +26,16 @@ locals {
     "pod-security.kubernetes.io/warn"  = "baseline"
   }
 
-  loki_output = {
-    enabled       = var.core_services_config.loki.enabled
-    host          = var.core_services_config.loki.enabled ? module.loki[0].host : ""
-    port          = var.core_services_config.loki.enabled ? module.loki[0].port : -1
-    node_logs     = var.core_services_config.loki.node_logs
-    workload_logs = true
+  loki_nodes_output = {
+    enabled = var.logging.nodes.loki.enabled
+    host    = var.logging.nodes.loki.enabled ? module.loki[0].host : null
+    port    = var.logging.nodes.loki.enabled ? module.loki[0].port : null
+  }
+
+  loki_workloads_output = {
+    enabled = var.logging.workloads.loki.enabled
+    host    = var.logging.workloads.loki.enabled ? module.loki[0].host : null
+    port    = var.logging.workloads.loki.enabled ? module.loki[0].port : null
   }
 
   azure_storage_nodes_output = {